Bug fixes and changes:
Added regular expressions for pattern matching in SubmissionErrorsListener, abstracted the pattern generation and matches to Container.

Added SQLMaintanceTask to execute SQL statements contianed in a file pointed to by Project.xml.

Some bugfixes to ArchiveTask, added ParseDartTimeStamp to test date time stamps from the command line.



git-svn-id: http://svn.na-mic.org/Dart/trunk@424 b2c14c61-d7ef-0310-8715-841e9ab387fe

+++ Source/dart/server/task/DeleteDataTask.java
@@ -46,6 +46,10 @@ public class DeleteDataTask implements Task {
       } else {
         logger.debug ( "Deleting result data: " + ResultValue );
         File file = new File ( project.getDataDirectory() + File.separator + ResultValue );
+        if ( !file.exists() ) {
+          logger.error ( project.getTitle() + ": File does not exist: " + file.getPath() );
+          return;
+        }
         if ( !file.delete() ) {
           logger.error ( "Failed to delete file: " + file.getPath() );
           return;
